Moved Dependency graph Folder to Component Folder
