+++ src/app/component/mapping/mapping.component.html
@@ -101,7 +101,7 @@
     <tr mat-row *matRowDef="let row; columns: displayedColumns;"></tr>
   </table>
 
-  <table id="excel-table" >
+  <table id="excel-table" class="hide">
     <tr>
       <th>Dimension</th>
       <th>Sub Dimension</th>
